Make documentation and file changes
