+++ docs/source/conf.py
@@ -47,7 +47,8 @@ sys.path.append(os.path.join(SAGE_SRC, "sage_setup", "docbuild", "ext"))
 # ones.
 extensions = [
     #'sphinx.ext.autodoc',
-    'sage_autodoc',
+    #'sage_autodoc',
+    'sage_docbuild.ext.sage_autodoc',
     'sage_package.sphinx',
     'sphinx.ext.doctest',
     'sphinx.ext.coverage',
@@ -232,7 +233,7 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-  ('index', package_name + '.tex', u'Documentation of ' + unicode(package_name),
+  ('index', package_name + '.tex', u'Documentation of ' + str(package_name),
    authors, 'manual'),
 ]
 
@@ -262,7 +263,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    ('index', package_name, unicode(package_name) + u" documentation",
+    ('index', package_name, str(package_name) + u" documentation",
      [authors], 1)
 ]
 
@@ -276,7 +277,7 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-  ('index', package_name, unicode(package_name) + u" documentation",
+  ('index', package_name, str(package_name) + u" documentation",
    authors, package_name, project,
    'Miscellaneous'),
 ]
@@ -322,7 +323,7 @@ if (os.environ.get('SAGE_DOC_MATHJAX', 'no') != 'no'
     exclude_patterns=['**/'+os.path.join(mathjax_relative, i) for i in ('docs', 'README*', 'test',
                                                                         'unpacked', 'LICENSE')]
 else:
-     extensions.append('sphinx.ext.pngmath')
+     extensions.append('sphinx.ext.imgmath')
 
 # This is to make the verbatim font smaller;
 # Verbatim environment is not breaking long lines
