--- docs/source/oriented_matroids/hyperplane_arrangement_oriented_matroid.rst
@@ -1,9 +1,0 @@
-.. nodoctest
-
-Hyperplane Arrangement Oriented Matroids
-========================================
-
-.. automodule:: oriented_matroids.hyperplane_arrangement_oriented_matroid
-    :members:
-    :undoc-members:
-    :show-inheritance:
