--- docs/source/oriented_matroids/signed_vector_element.rst
@@ -1,9 +1,0 @@
-.. nodoctest
-
-Signed Vectors
-==============
-
-.. automodule:: oriented_matroids.signed_vector_element
-    :members:
-    :undoc-members:
-    :show-inheritance:
