+++ makefile
@@ -4,12 +4,12 @@
 PACKAGE = oriented_matroids
 
 # change to your sage command if needed
-SAGE = /sage/sagedev/sage
+SAGE = /sage/sage/sage
 
 all: install test
 
 install:
-	$(SAGE) -pip install --upgrade --no-index -v .
+	$(SAGE) -pip install --upgrade --no-index --no-build-isolation -v .
 
 uninstall:
 	$(SAGE) -pip uninstall $(PACKAGE)
