+++ oriented_matroids/__init__.py
@@ -2,6 +2,6 @@ from __future__ import absolute_import
 from .oriented_matroid import OrientedMatroid
 from .abstract_oriented_matroid import AbstractOrientedMatroid
 
-from sage.misc.lazy_import import lazy_import
-#lazy_import('sage.matroids.oriented_matroids',
+# from sage.misc.lazy_import import lazy_import
+# lazy_import('sage.matroids.oriented_matroids',
 #            'oriented_matroids_catalog', 'oriented_matroids')
