+++ oriented_matroids/circuit_oriented_matroid.py
@@ -10,7 +10,7 @@ AUTHORS:
 """
 
 ##############################################################################
-#       Copyright (C) 2019 Aram Dermenjian <aram.dermenjian at gmail.com>
+#       Copyright (C) 2019 Aram Dermenjian <aram.dermenjian.math at gmail.com>
 #
 #  Distributed under the terms of the GNU General Public License (GPL)
 #
@@ -149,7 +149,7 @@ class CircuitOrientedMatroid(AbstractOrientedMatroid):
             Traceback (most recent call last):
             ...
             ValueError: Every element needs an opposite
-            
+
             sage: C5 = [((1,),(3,),(2,)), ((1,2),(3,),(4,)), ((3,),(1,),(2,)), ((3,),(1,2),(4,))]
             sage: OrientedMatroid(C5,key='circuit')
             Traceback (most recent call last):
