+++ oriented_matroids/oriented_matroids_catalog.py
@@ -8,4 +8,4 @@ A module containing constructors for several common oriented matroids
 # Do not add code to this file, only imports.
 
 # user-accessible:
-# from sage.matroids.catalog import ...
+# from oriented_matroids.catalog import ...
