+++ oriented_matroids/real_hyperplane_arrangement_oriented_matroid.py
@@ -10,7 +10,7 @@ AUTHORS:
 """
 
 ##############################################################################
-#       Copyright (C) 2019 Aram Dermenjian <aram.dermenjian at gmail.com>
+#       Copyright (C) 2019 Aram Dermenjian <aram.dermenjian.math at gmail.com>
 #
 #  Distributed under the terms of the GNU General Public License (GPL)
 #
@@ -18,10 +18,9 @@ AUTHORS:
 #
 #                  http://www.gnu.org/licenses/
 ##############################################################################
-
-
 from oriented_matroids.covector_oriented_matroid import CovectorOrientedMatroid
 from sage.categories.sets_cat import Sets
+
 
 class RealHyperplaneArrangementOrientedMatroid(CovectorOrientedMatroid):
     r"""
@@ -60,7 +59,8 @@ class RealHyperplaneArrangementOrientedMatroid(CovectorOrientedMatroid):
     .. SEEALSO::
 
         :class:`oriented_matroids.oriented_matroid.OrientedMatroid`
-        :class:`oriented_matroids.oriented_matroids_category.OrientedMatroids`
+        :class:`oriented_matroids.abstract_oriented_matroid.AbstractOrientedMatroid`
+        :class:`sage.geometry.hyperplane_arrangement.arrangement.HyperplaneArrangementElement`
     """
 
     @staticmethod
@@ -108,8 +108,7 @@ class RealHyperplaneArrangementOrientedMatroid(CovectorOrientedMatroid):
         """
 
         if not self.arrangement().is_central():
-            raise ValueError(
-                "Hyperplane arrangements must be central to be an oriented matroid.")
+            raise ValueError("Hyperplane arrangements must be central to be an oriented matroid.")
 
         return True
 
@@ -140,7 +139,6 @@ class RealHyperplaneArrangementOrientedMatroid(CovectorOrientedMatroid):
             sage: G = Graph({1:[2,4],2:[3,4,5],3:[4,6,8],4:[7],5:[8]})
             sage: A = hyperplane_arrangements.graphical(G)
             sage: H = [A.hyperplanes()[i] for i in range(2,5)]
-            sage: M = OrientedMatroid(A)
             sage: M = OrientedMatroid(A); M
             Hyperplane arrangement oriented matroid of rank 7
             sage: M2 = M.deletion(H); M2
