+++ oriented_matroids/vector_oriented_matroid.py
@@ -10,7 +10,7 @@ AUTHORS:
 """
 
 ##############################################################################
-#       Copyright (C) 2018 Aram Dermenjian <aram.dermenjian at gmail.com>
+#       Copyright (C) 2018 Aram Dermenjian <aram.dermenjian.math at gmail.com>
 #
 #  Distributed under the terms of the GNU General Public License (GPL)
 #
@@ -48,14 +48,14 @@ class VectorOrientedMatroid(AbstractOrientedMatroid):
 
     INPUT:
 
-    - ``data`` -- a tuple containing SigneVectorElement elements or data
-      that can be used to construct :class:`SignedVectorElement` elements
+    - ``data`` -- a tuple containing SignedSubsetElement elements or data
+      that can be used to construct :class:`SignedSubsetElement` elements
     - ``goundset`` -- (default: ``None``) is the groundset for the
       data. If not provided, we grab the data from the signed subsets.
 
     EXAMPLES::
 
-        sage: from oriented_matroids import OrientedMatroid
+        sage: from oriented_matroids.oriented_matroid import OrientedMatroid
         sage: M = OrientedMatroid([[1],[-1],[0]], key='vector'); M
         Vector oriented matroid of rank 0
         sage: M.groundset()
@@ -70,7 +70,8 @@ class VectorOrientedMatroid(AbstractOrientedMatroid):
     .. SEEALSO::
 
         - :class:`oriented_matroids.oriented_matroid.OrientedMatroid`
-        - :class:`oriented_matroids.oriented_matroids_category.OrientedMatroids`
+        - :class:`oriented_matroids.abstract_oriented_matroid.AbstractOrientedMatroid`
+        - :class:`oriented_matroids.signed_subset_element.SignedSubsetElement`
     """
     @staticmethod
     def __classcall__(cls, data, groundset=None, category=None):
@@ -113,11 +114,11 @@ class VectorOrientedMatroid(AbstractOrientedMatroid):
 
     def is_valid(self):
         """
-        Returns whether our circuits satisfy the circuit axioms.
+        Return whether our vectors satisfy the vector axioms.
 
         EXAMPLES::
 
-            sage: from oriented_matroids import OrientedMatroid
+            sage: from oriented_matroids.oriented_matroid import OrientedMatroid
             sage: V2 = [[1,1]]
             sage: OrientedMatroid(V2, key='vector')
             Traceback (most recent call last):
@@ -186,7 +187,7 @@ class VectorOrientedMatroid(AbstractOrientedMatroid):
 
         EXAMPLES::
 
-            sage: from oriented_matroids import OrientedMatroid
+            sage: from oriented_matroids.oriented_matroid import OrientedMatroid
             sage: V = [[1,1],[-1,-1],[0,0]]
             sage: M = OrientedMatroid(V, key='vector'); M
             Vector oriented matroid of rank 1
@@ -197,21 +198,6 @@ class VectorOrientedMatroid(AbstractOrientedMatroid):
         except ValueError:
             rep = "Vector oriented matroid"
         return rep
-
-    def circuits(self):
-        """
-        Return the circuits.
-
-        Given a vector oriented matroid, the set of circuits is the set
-        `Min(V)` which denotes the set of inclusion-minimal (nonempty) signed
-        subsets.
-        """
-        from sage.combinat.posets.posets import Poset
-        from oriented_matroids import OrientedMatroid
-        # remove 0
-        vecs = [v for v in self.vectors() if not v.is_zero()]
-        P = Poset([vecs, lambda x,y: x.is_restriction_of(y)])
-        return P.minimal_elements()
 
     def matroid(self):
         r"""
@@ -224,18 +210,18 @@ class VectorOrientedMatroid(AbstractOrientedMatroid):
 
         EXAMPLES::
 
-            sage: from oriented_matroids import OrientedMatroid
+            sage: from oriented_matroids.oriented_matroid import OrientedMatroid
             sage: V = [[1,1],[-1,-1],[0,0]]
             sage: M = OrientedMatroid(V, key='vector')
             sage: M.matroid()
             Matroid of rank 1 on 2 elements with 2 bases
 
         """
-        circOM = self.to_circuit()
+        circOM = self.convert_to('circuit')
         return circOM.matroid()
 
     def circuits(self):
-        """
+        r"""
         Return the circuits.
 
         Given a vector oriented matroid, the set of circuits is the set
@@ -243,7 +229,8 @@ class VectorOrientedMatroid(AbstractOrientedMatroid):
         subsets.
 
         EXAMPLES::
-            sage: from oriented_matroids import OrientedMatroid
+
+            sage: from oriented_matroids.oriented_matroid import OrientedMatroid
             sage: M = OrientedMatroid([[1],[-1],[0]], key='vector')
             sage: M.circuits()
             [+: 0
@@ -253,9 +240,11 @@ class VectorOrientedMatroid(AbstractOrientedMatroid):
              -: 0
              0: ]
         """
+        if hasattr(self, "_circuits"):
+            return self._circuits
         from sage.combinat.posets.posets import Poset
-        from oriented_matroids import OrientedMatroid
         # remove 0
         vecs = [v for v in self.vectors() if not v.is_zero()]
-        P = Poset([vecs, lambda x,y: x.is_restriction_of(y)])
-        return P.minimal_elements()
+        P = Poset([vecs, lambda x, y: x.is_restriction_of(y)])
+        self._circuits = P.minimal_elements()
+        return self._circuits
